Redesign the render API
These improvements allow us to add more configuration to the rendering
process in the future. It also prevents need for a proliferation of
rendering functions. The new public `Renderer` struct is created using
one of the three functions: `render`, `render_from`, or
`render_from_fn`. And then the renderer can be consumed using
`.to_string()` or `.to_writer(..)`.